Added a new design.
The color palette is inspired on "old west" colors.
The font is somewhat vintage.
The background now has a gradient resembling a sunset.
The table is now round.
Styled edit and share buttons.
Remake the cards.[this is not yet finished]
There are a few glowing buttons.
