+++ client/public/index.html
@@ -27,6 +27,9 @@
 
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons"
       rel="stylesheet">
+      <link rel="preconnect" href="https://fonts.googleapis.com">
+      <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+      <link href="https://fonts.googleapis.com/css2?family=DM+Serif+Display&family=PT+Sans:wght@400;700&family=Playfair+Display&display=swap" rel="stylesheet"> 
     <title>React App</title>
   </head>
   <body>
